feat: filter supplier debt account

+++ debt-service/src/main/java/com/ttdat/debtservice/api/dto/common/DebtAccountDTO.java
@@ -29,7 +29,7 @@ public class DebtAccountDTO {
 
     BigDecimal remainingAmount;
 
-    Double interestRate;
+    BigDecimal interestRate;
 
     LocalDate dueDate;
