+++ debt-service/src/main/java/com/ttdat/debtservice/api/dto/response/PartyDebtAccount.java
@@ -25,7 +25,7 @@ public class PartyDebtAccount {
 
     BigDecimal remainingAmount;
 
-    Double interestRate;
+    BigDecimal interestRate;
 
     LocalDate dueDate;
 }
